feature(conditions): move compound condition to base namespace
