+++ packages/taco/src/conditions/index.ts
@@ -4,7 +4,7 @@ import * as predefined from './predefined';
 export {
   CompoundConditionProps,
   CompoundConditionType,
-} from './compound-condition';
+} from './base/compound-condition';
 export { Condition, ConditionProps } from './condition';
 export { ConditionExpression, ConditionExpressionJSON } from './condition-expr';
 export { ConditionContext, CustomContextParam } from './context';
